Decode: archive offsets are 64-bit on every build, not size_t
`nz-re` x86 answers "Out of memory!" to a 4.6 GB archive the original reads
without trouble. Reproducing it turned up TWO independent blockers, and neither
one alone explains the refusal:

  * the archive is mapped whole, and 4.6 GB does not fit in the ~3 GB of user
    address space a 32-bit process gets;
  * every offset into the archive was `std::size_t`, which on a 32-bit build is
    32 bits -- so a position past 4 GB cannot even be REPRESENTED, and windowing
    the mapping would not have helped by itself.

This is the second of those. `ArcPos` (an explicit `std::uint64_t`) now names a
position or length inside the archive, and `ByteView`, `ReadLegacyVarint`,
`ReadLegacyTableSpan(Flexible)`, `TryAssembleStoredBlocks`, `DecodeLzpfMember`,
`nzr::derr`'s failure position and every offset local in the parser use it.

The size of the job was measured, not estimated: widening `ByteView` in a scratch
copy and compiling with `-m32` made the compiler enumerate the sites -- 40
warnings and 1 error, not the 335 uses of the view I had feared. The conversion
was then driven to zero errors AND zero archive-offset narrowings under
`-m32 -Wconversion`.

What that shook out is the distinction the old type hid: offsets into the ARCHIVE
(now 64-bit everywhere) against offsets into an in-memory WINDOW or output buffer
(still `size_t`, with the narrowing made explicit and commented where a value is
bounded by one block). I widened `block_start_in_window` at first, which is a
window position and not an archive one; the warnings caught it.

On a 64-bit build this is a no-op -- `size_t` is already 64 bits there. The point
is the 32-bit build, so that is where it was checked: the full x86 binary builds
clean and decodes 96 of 96 archives byte-exact with zero bridge.

64-bit regression, all level with the previous commit: unit tests unchanged, 96
native decodes byte-exact, damaged single-container 47/48 files and 34/48
messages, damaged parallel 35/35, console parity one pre-existing difference, the
encode oracle 133/135 byte-identical (the two are -cO and -cc, still unported;
the extra console entry is quirk 15's timing-dependent progress redraw count).

NOT fixed yet, and deliberately so: x86 still refuses the 4.6 GB archive, because
the mapping is still whole-file. That is the other half -- the ~85 sites that take
a raw pointer into the map have to ask for a resident RANGE instead. The per-chunk
decode from the previous commit already consumes the archive in bounded ranges,
which is the shape that work needs.

+++ include/nz_decode_error.h
@@ -20,7 +20,10 @@ namespace derr {
 struct State {
     std::uint32_t code = 0;       // the original's status (0 = none recorded)
     std::uint32_t fatal_id = 0;   // "Internal error: <id>!" assertion id (0 = none)
-    std::size_t input_pos = 0;    // input offset of the failing block, when known
+    // Input offset of the failing block, when known. 64-bit, not size_t: a
+    // 32-bit build still has to name a position inside an archive larger than
+    // its own address space.
+    std::uint64_t input_pos = 0;
     bool has_pos = false;
     bool parallel = false;        // failure of one worker stream of a parallel container
     std::size_t slot = 0;         // that stream's slot (the original's low report byte)
@@ -30,14 +33,14 @@ inline void Clear() { t_state = State{}; }
 inline void Set(std::uint32_t code) {
     if (t_state.code == 0u && t_state.fatal_id == 0u) t_state.code = code;
 }
-inline void SetAt(std::uint32_t code, std::size_t input_pos) {
+inline void SetAt(std::uint32_t code, std::uint64_t input_pos) {
     if (t_state.code == 0u && t_state.fatal_id == 0u) { t_state.code = code; t_state.input_pos = input_pos; t_state.has_pos = true; }
 }
 inline void Fatal(std::uint32_t id) { if (t_state.fatal_id == 0u) t_state.fatal_id = id; }
 // A failure with no status of its own (the engine declined) still has a PLACE:
 // record it so the plain-vs-shifted rule can see whether another record of the
 // stream lay ahead. Never overrides a position a status already recorded.
-inline void SetPos(std::size_t input_pos) {
+inline void SetPos(std::uint64_t input_pos) {
     if (!t_state.has_pos) { t_state.input_pos = input_pos; t_state.has_pos = true; }
 }
 inline const State& Current() { return t_state; }
